Refactored files to match convention
