+++ swarmdjango/core/tests.py
@@ -1,3 +1,3 @@
 from django.test import TestCase
 
-# Create your unit_tests here.
+# Create your unitTests here.
